Update transforms.py
I have a little bit change, here "In addition, The minimum size of the cropped image is [width * retain_ratio[0], height * retain_ratio[1]". I corrected the error to "In addition, The minimum size of the cropped image is [width * retain_ratio[0], height * retain_ratio[1]]",  In other words，I add a "]" to " retain_ratio[1]".