Update text_to_sam_clip.py (#3299)
The argument passed to the TabbedInterface of gradio is outdated. Now, instead of title, it is theme. Because of this running the command:
python scripts/text_to_sam_clip.py --model-type vit_h

fails with the erro: TypeError: TabbedInterface.__init__() got an unexpected keyword argument 'title'.

Hence this commit, resolved this error by updating the argument passed.